+++ Plans/website-charts/stable/mosquitto/CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 
+## [mosquitto-6.0.22](https://github.com/truecharts/apps/compare/mosquitto-6.0.21...mosquitto-6.0.22) (2022-07-20)
+
+
+
+
+## [mosquitto-6.0.20](https://github.com/truecharts/apps/compare/mosquitto-6.0.19...mosquitto-6.0.20) (2022-07-19)
+
+### Chore
+
+- Update Readme and Description on SCALE
+
+
+
 ## [mosquitto-6.0.19](https://github.com/truecharts/apps/compare/mosquitto-6.0.17...mosquitto-6.0.19) (2022-07-19)
 
 ### Feat